tema3/image-processing-app/appengine/requirements.txt
Flask
google-cloud-firestore
google-cloud-storage
google-auth
gunicorn